Transaction id is now an 8 byte integer instead of 4

+++ src/main/java/ru/magzook/sellersrestservice/controller/TransactionController.java
@@ -25,7 +25,7 @@ public class TransactionController {
     }
 
     @GetMapping("/transactions/{id}")
-    public TransactionWithSellerDto findById(@PathVariable int id) {
+    public TransactionWithSellerDto findById(@PathVariable long id) {
         return transactionService.findByIdFetchSeller(id);
     }
